Rename auth hook file to useAuth
